requirements.txt
# core ML Libraries
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
xgboost>=1.5.0 
torch>=1.9.0

# visualisation
matplotlib>=3.4.0
seaborn>=0.11.0

# configuration
pyyaml>=5.4.0

# utilities
joblib>=1.0.0
